pyproject.toml
[project]
name = "agent-notes"
version = "0.2.0"
description = "Agentic Memory via Git Notes"
readme = "README.md"
authors = [
    { name = "Dallas Pool", email = "codeninja@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.2",
    "gitpython>=3.1.46",
    "pydantic>=2.12.5",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.urls]
Homepage = "https://github.com/codeninja/agent-notes"
Repository = "https://github.com/codeninja/agent-notes"

[project.scripts]
agent-notes = "agent_notes.main:app"
agent-notes-dx = "agent_notes.dx:app"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
